### 2025-05-29 #8313 @editorialbot
[PRE REVIEW]: SkyRank: a unified toolkit for Skyline ranking, benchmarking and visualization
**Submitting author:** @MickaelMartinNevot (<a href="http://orcid.org/0009-0004-7893-3449">Mickaël Martin Nevot</a>)
**Repository:** https://github.com/Turlure-Nael-23018992/SkyRank
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
