### 2025-06-01 #8313 @MickaelMartinNevot
> thanks for your submission to JOSS. You software looks to be just a little under two months old. Can you tell me a bit more about the development history for this software? Thanks.

The Skyline ranking problematics such as those underlying the SkyRank algorithmic platform was studied in first place, in french, in "Extraction multidimensionnelle et multicritère de données émergentes pour l'équilibrage de jeux vidéo" (Martin Nevot 2023 - [https://theses.fr/2023AIXM0361](https://theses.fr/2023AIXM0361)) thesis, and "Classement d'objets Skylines dans les bases de données" (Martin Nevot et al 2024 - https://arxiv.org/abs/2411.02013) conference.

We quickly came up with the idea and desire to develop an open-source algorithmic platform solution to offer implementations of methods from the literature or those proposed as part of our research work. This submission concerns a toolbox offering different proven methods for ranking Skyline objects in multiple-criteria decision analysis, the possibility of comparing them, and the synthesis as a visualization of the results in Latex/TikZ. The platform is mutimodal, it can be used as is, or easily as part of another project.
