### 2025-06-02 #8313 @crvernon
Thanks @MickaelMartinNevot for the additional information.  I am going to run this through scope query with our larger editorial board to ensure that it meets our requirements for substantial scholarly effort.  I'll be back in touch ASAP.
