### 2025-07-04 #8313 @abhishektiwari
@MickaelMartinNevot

- [ ] Please add guidelines for contribution `CONTRIBUTING.md`. See [a generator here](https://contributing.md/generator/).
- [ ] Can you please clarify if scope of this submission? Are SkyRank client and [Git Submodules](https://github.com/Turlure-Nael-23018992/BBS-Python-3.x-) part of submission along with SkyRank?
- [ ] Can you clarify use of Git submodules as mentioned above? Is there a reasons why they can not be installed as package via PIP+Git?
- [ ] `docs` folder is autogenerated. If so where is source documentation?
