### 2025-07-05 #8313 @MickaelMartinNevot
- `CONTRIBUTING.md` and `CODE_OF_CONDUCT.md`files directly added on repository root.
 - We can focus on SkyRank client only. Git Submodules are relevant and integral part of the projet, but there are side issues.
 - Git submodules are independant semantic projects. A recursive project clone automatically install every Git submodules. Don't hesitate to take a look at `README.rst` file at repository root for further information on this topic.
 - True, `docs` folder is autogenerated. The source documentation is included directly in the source code before each function in each file.
