### 2025-07-06 #8313 @abhishektiwari
@MickaelMartinNevot Given following dependency architecture, all 4 repositories have to be in scope. I also think each repository needs to have appropriate License and Contribution guidelines. Can you clarify what side issues you see covering all 4 as part of review?

```
SkyRank-Client
  └── SkyRank
        └── BBS-Python-3.x-
              └── R-Tree-Python-3.x-
```

Since the documentation is auto-generated from docstrings, I recommend adding the generated HTML files to .gitignore to keep them out of version control. Instead, you should implement a GitHub Actions workflow to automatically build and deploy the documentation.

While this doesn't need to be implemented immediately, I'd like to see this automation added as part of the review process. The current approach of committing generated files is creating noise in git history—the repository currently shows 90,720 lines added and 30,526 removed, which makes it difficult for reviewers to focus on the actual code changes.

Setting up automated documentation builds will help maintain a cleaner commit history and make code reviews more manageable.
