### 2025-07-12 #8313 @MickaelMartinNevot
For this kind of work, we are really not sure. Maybe, without COI of course, Jérôme Darmont ([jerome.darmont@univ-lyon2.fr](mailto:jerome.darmont@univ-lyon2.fr)), Fadila Bentayeb ([fadila.bentayeb@univ-lyon2.fr](mailto:fadila.bentayeb@univ-lyon2.fr)) and Noël Novelli ([noel.novelli@lis-lab.fr](noel.novelli@lis-lab.fr)).
