### 2025-07-24 #8313 @abhishektiwari
👋 @tommyod @BillKasapidis @hanneseder-net @rnd-forests, would any of you be willing to review this submission for JOSS?  The submission under consideration is "SkyRank: a unified toolkit for Skyline ranking, benchmarking and visualization".

Based on your experience with Skyline Operator/Queries/Ranking, we think you might be able to provide a great review of this submission.

Thanks
Abhi
