### 2025-08-08 #8313 @abhishektiwari
@shrishar Saw your notes on other pre-review thread. Looking at your experience with data engineering and distributed systems, I think [this submission](https://github.com/openjournals/joss-reviews/issues/8313) will be a better fit for you to review.
