### 2025-08-15 #8313 @abhishektiwari
Thanks for getting back to me @MNISAR. Happy to clarify scope of ask here. Journal of Open Source Software (JOSS) is a developer friendly, open access journal for research software packages. Ask here is to review paper “SkyRank: a unified toolkit for Skyline ranking, benchmarking and visualization” and associated software.

We follow a checklist-driven reviews here in GitHub issues. See review guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
