### 2025-08-16 #8313 @Pranshu258
@abhishektiwari I have worked on 2 projects involving skyline queries before. Although that was several years ago, I can take a shot at reviewing this submission. Can you please clarify the specifics, as in, where do we submit the reviews, and is there format in which the review must be written?
