### 2025-09-21 #8834 @abhishektiwari
Thanks @Pranshu258

@MickaelMartinNevot Please clarify the authorship particularly your contribution to software and paper. See JOSS Guidance [here](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship). Let me know if any questions.

> [@abhishektiwari](https://github.com/abhishektiwari) - "Contribution and authorship: Has the submitting author ([@MickaelMartinNevot](https://github.com/MickaelMartinNevot)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?"
>
> Looking at the commit history, the submitting author does not have the most contributions. How do we proceed in this case?
>
> <img alt="Image" width="918" height="680" src="https://private-user-images.githubusercontent.com/7602984/489397205-de8898d6-bf9c-4d40-92d4-5e95c85cef60.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NTg0NjE0ODMsIm5iZiI6MTc1ODQ2MTE4MywicGF0aCI6Ii83NjAyOTg0LzQ4OTM5NzIwNS1kZTg4OThkNi1iZjljLTRkNDAtOTJkNC01ZTk1Yzg1Y2VmNjAucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDkyMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTA5MjFUMTMyNjIzWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ODNlZDUzMTcwYTgxMWNlN2E5ZTRjN2Q2ZjQwMjdiMjMyY2Q0MTQwMzI4MTBlOGFmZDQ1NzBhZmEzZGUwNjBjYiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.umHzMDEllzvN7P4KZ_nbBqRpNL2GkI_U6_dYyfU0SiY">
