### 2025-09-21 #8834 @abhishektiwari
@Pranshu258 - Please let me know if you still have questions about authorship based on above response. See JOSS Guidance [here](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship).
