### 2025-10-28 Turlure-Nael-23018992/SkyRank#1 @abhishektiwari
[JOSS] Installation challenges with Ubuntu 23
Reviewer's [feedback on the installation challenges with  Ubuntu 23](https://github.com/openjournals/joss-reviews/issues/8834#issuecomment-3420032929)

> [trying on ubuntu23 - failed as it needed manual setup] 🔴
>
> reason:
>
> Installation for linux (ubuntu 23) is not super easy. I had to go through multiple troubles to get the setup.
> python 3.8 is not directly available in Linux (only Win & mac): [AI] include instruction for linux
> make and install version 3.8 manually (stressful)
> openssl configuration needed in addition

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8834)
