### 2025-11-01 Turlure-Nael-23018992/SkyRank#2 @proy-11
[JOSS] Unit Tests failure
Please fix the failing unit tests.
```
# Success
python -m unittest Algorithms.Tests.RankSkyTest -v  # 8/8 tests pass

# Failing
python -m unittest Algorithms.Tests.CoskyAlgoTest -v     # FileNotFoundError: '../Datas/RBig.json'
python -m unittest Algorithms.Tests.DpIdpDhTest -v       # FileNotFoundError: '../Datas/RBig.json'
python -m unittest Algorithms.Tests.DeepSkyTest -v       # FileNotFoundError: '../Datas/Rtuples8.json'
```
