### 2025-11-01 Turlure-Nael-23018992/SkyRank#3 @proy-11
[JOSS] .gitignore issue - virtual env not excluded
The installation script `install.sh` creates the virtual environment, and installs the dependencies.
However, `.venv/` folder is not added to `.gitignore`,  causing virtual env with all dependencies to be tracked in git.
Please consider including more comprehensive python exclusions:
```
.venv/
__pycache__/
*.pyc
*.pyo
*.egg-info/
```
