### 2025-11-01 Turlure-Nael-23018992/SkyRank#4 @proy-11
[JOSS] Dependency management issue
`pyproject.toml` and `requirements.txt` both are present, with conflicting versions of `numpy`, `matplotlib`, and `scikit-learn`.
- The installation script (`install.sh`) uses `pyproject.toml` to install dependencies, so `requirements.txt` appears unused.
- This may create confusion about which file governs dependencies, maintenance burden and potential compatibility issues.
- Please remove the redundant dependency file, or document the purpose and align versions if both required.
