### 2025-11-01 #8834 @proy-11
> Hi [@proy-11](https://github.com/proy-11) How are you going with review? Can you update us with your progress?

Hi @abhishektiwari
Thank you for checking-in. Please find an update on my progress below:

- Installed and tested SkyRank on macOS.
- Verified GUI, basic functionality, and tests.
- Identified these issues:
    - [SkyRank/issues/2](https://github.com/Turlure-Nael-23018992/SkyRank/issues/2)
    - [SkyRank/issues/3](https://github.com/Turlure-Nael-23018992/SkyRank/issues/3)
    - [SkyRank/issues/4](https://github.com/Turlure-Nael-23018992/SkyRank/issues/4)

Next Steps: I'm going through the paper and documentation. I'll update the checklist shortly.
