### 2025-11-04 f39971c
update failing unit tests for the isssue #2
