+++ Algorithms/Tests/DeepSkyTest.py
@@ -4,7 +4,6 @@ import os
 import sys
 
 from Algorithms.CoskyAlgorithme import CoskyAlgorithme
-from Algorithms.SkyIR import SkyIR
 from Utils.DataModifier.DataNormalizerDeepSky import DataNormalizerDeepSky
 from Utils.DataModifier.JsonUtils import readJson
 from Algorithms.CoskySql import CoskySQL
@@ -14,18 +13,21 @@ from Algorithms.RankSky import RankSky
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
 
-
 class DeepSkyTest(unittest.TestCase):
     def setUp(self):
-        self.rTuples8 = readJson("../Datas/Rtuples8.json")
+        base_dir = os.path.dirname(__file__)
+        rTuple = os.path.abspath(os.path.join(base_dir, "..", "Datas", "rTuples8.json"))
+        self.rTuples8 = readJson(rTuple)
         self.k = 7
-        self.dataNorm = DataNormalizerDeepSky(self.rTuples8, "../../Assets/DeepSkyTest.db")
+        rbig_path = os.path.abspath(os.path.join(base_dir, "..", "Datas", "DeepSkyTest.db"))
+        self.path = rbig_path
+        self.dataNorm = DataNormalizerDeepSky(self.rTuples8, self.path)
 
     def testDeepSkyCoskySql(self):
         """
         Test the DeepSky algorithm with CoskySql.
         """
-        deepSky = DeepSky("../../Assets/DeepSkyTest.db", self.k, CoskySQL)
+        deepSky = DeepSky(self.path, self.k, CoskySQL)
         result = deepSky.topK
         self.assertEqual(self.k, len(result), "The result should contain k elements.")
 
@@ -33,23 +35,16 @@ class DeepSkyTest(unittest.TestCase):
         """
         Test the DeepSky algorithm with CoskyAlgo.
         """
-        deepSky = DeepSky("../../Assets/DeepSkyTest.db", self.k, CoskyAlgorithme)
+        deepSky = DeepSky(self.path, self.k, CoskyAlgorithme)
         result = deepSky.topK
         self.assertEqual(self.k, len(result), "The result should contain k elements.")
 
     '''def testDeepSkyDpIdpDh(self):
+        DpIdp with deepSky currently in dev !
         """
         Test the DeepSky algorithm with DpIdpDh.
         """
-        deepSky = DeepSky("../../Assets/DeepSkyTest.db", self.k, DpIdpDh)
-        result = deepSky.topK
-        self.assertEqual(self.k, len(result), "The result should contain k elements.")'''
-
-    '''def testDeepSkySkyIR(self):
-        """
-        Test the DeepSky algorithm with SkyIR.
-        """
-        deepSky = DeepSky("../../Assets/DeepSkyTest.db", self.k, SkyIR)
+        deepSky = DeepSky("../../Algorithms/Datas/DeepSkyTest.db", self.k, DpIdpDh)
         result = deepSky.topK
         self.assertEqual(self.k, len(result), "The result should contain k elements.")'''
 
@@ -57,7 +52,7 @@ class DeepSkyTest(unittest.TestCase):
         """
         Test the DeepSky algorithm with RankSky.
         """
-        deepSky = DeepSky("../../Assets/DeepSkyTest.db", self.k, RankSky)
+        deepSky = DeepSky(self.path, self.k, RankSky)
         result = deepSky.topK
         self.assertEqual(self.k, len(result), "The result should contain k elements.")
 
@@ -65,36 +60,14 @@ class DeepSkyTest(unittest.TestCase):
         """
[…]
