### 2025-11-04 f03621c
added other extensions in .gitignore for issue #3
