### 2025-11-04 d40a4e2
removed requirements.txt from the github to have pyproject.toml as single dependency manager for Dependency management issue #4
