+++ requirements.txt
@@ -1,13 +1,0 @@
-pip~=25.0.1
-numpy~=2.2.5
-PyQt5~=5.15.11
-matplotlib~=3.10.3
-scikit-learn~=1.6.1
-scipy~=1.15.3
-colorama~=0.4.6
-humanize~=4.12.3
-six~=1.17.0
-pillow~=11.2.1
-psutil
-GPUtil
-tabulate
