### 2025-11-05 3f3a993
update README.rst with Python 3.8 installation instructions for Ubuntu 23+
