+++ README.rst
@@ -16,7 +16,7 @@ It is designed for researchers, students, and practitioners interested in prefer
 - Git installed
 - Python 3.8 installed (https://www.python.org/downloads/release/python-380/)
 - Internet connection
-- (On Windows) Python launcher `py`
+- (On Windows) Python launcher ``py``
 
 📥 Clone the Repository with Submodules
 ---------------------------------------
@@ -32,11 +32,12 @@ To clone the full project with all submodules in one step:
 
     git clone --recurse-submodules https://github.com/Turlure-Nael-23018992/SkyRank.git
 
-If you already cloned it without `--recurse-submodules`, run:
+If you already cloned it without ``--recurse-submodules``, run:
 
 .. code-block:: bash
 
     git submodule update --init --recursive
+
 
 🚀 Installation on Linux/macOS
 ------------------------------
@@ -50,21 +51,64 @@ Run the provided shell script:
 
 This script will:
 
-- Use `python3.8` to create a virtual environment in `.venv`
+- Use ``python3.8`` to create a virtual environment in ``.venv``
 - Install all dependencies
 - Initialize Git submodules
-- Make the `skyrank-gui` and `skyrank-gui2` commands available inside the virtual environment
+- Make the ``skyrank-gui`` and ``skyrank-gui2`` commands available inside the virtual environment
+
+
+🐧 Installing on Ubuntu 23 and newer versions
+---------------------------------------------
+
+Ubuntu 23 and later no longer provide Python 3.8 in the default repositories.
+If you try to run ``install.sh``, you may see an error saying **“Python 3.8 not found.”**
+You have two simple options:
+
+**Option 1 — Install Python 3.8 from deadsnakes PPA**
+
+.. code-block:: bash
+
+    sudo apt update
+    sudo apt install -y software-properties-common
+    sudo add-apt-repository ppa:deadsnakes/ppa -y
+    sudo apt update
+    sudo apt install -y python3.8 python3.8-venv python3.8-dev python3.8-distutils
+
+Then create and activate a virtual environment:
+
+.. code-block:: bash
+
+    python3.8 -m venv .venv
+    source .venv/bin/activate
+    bash install.sh
+
+**Option 2 — Install Python 3.8 with pyenv (no system changes)**
+
+.. code-block:: bash
+
+    curl https://pyenv.run | bash
+    export PATH="$HOME/.pyenv/bin:$PATH"
+    eval "$(pyenv init -)"
+    eval "$(pyenv virtualenv-init -)"
+
+    pyenv install 3.8.18
+    pyenv local 3.8.18
+    python -m venv .venv
+    source .venv/bin/activate
+    bash install.sh
+
+
 
 🚀 Installation on Windows
 --------------------------
 
 SkyRank provides an installation script that does everything for you:
 
-1. **Checks for Python 3.8**
-2. **Creates `.venv` with CPython 3.8**
-3. **Installs all dependencies**
-4. **Initializes submodules**
-5. **Exposes CLI commands: `skyrank-gui`, `skyrank-gui2`**
+1. Checks for Python 3.8
+2. Creates ``.venv`` with CPython 3.8
+3. Installs all dependencies
+4. Initializes submodules
+5. Exposes CLI commands: ``skyrank-gui``, ``skyrank-gui2``
 
 To launch it:
 
@@ -73,6 +117,7 @@ To launch it:
     install.bat
 
 If you encounter a permissions issue, try running the terminal **as administrator**.
+
 
 🧪 Manual Setup (Alternative)
 -----------------------------
@@ -100,6 +145,7 @@ On **Windows**:
     python -m pip install -e .
     git submodule update --init --recursive
 
+
 🏁 Usage
 --------
 
[…]
