### 2025-11-05 Turlure-Nael-23018992/SkyRank#1 @MickaelMartinNevot
Until now, we have favored using Python 3.8 because of its long-term compatibility and stability, as well as because upstream dependencies still support it. We just added information about installing Python 3.8 on Ubuntu 23+ to the `readme` file. At the same time, we started a branch for the latest versions of Python. Although this introduces breaking changes that require significant refactoring, testing, and quality assurance (QA), we hope to offer such a version soon.
