### 2025-11-05 #8834 @MickaelMartinNevot
- [SkyRank/issues/1] (https://github.com/Turlure-Nael-23018992/SkyRank/issues/1): until now, we have favored using Python 3.8 because of its long-term compatibility and stability, as well as because upstream dependencies still support it. We just added information about installing Python 3.8 on Ubuntu 23+ to the `readme` file. At the same time, we started a branch for the latest versions of Python. Although this introduces breaking changes that require significant refactoring, testing, and quality assurance (QA), we hope to offer such a version soon.
 - [SkyRank/issues/2] (https://github.com/Turlure-Nael-23018992/SkyRank/issues/2): `DpIdpDh` and `CoskyAlgo` now only compare returned data sample points, not all values, which could potentially introduce rounding errors. The `DeepSkyTest` database path has been fixed. The DB has been recreated and moved.
 - [SkyRank/issues/3] (https://github.com/Turlure-Nael-23018992/SkyRank/issues/3): The recommended extensions and directories have now been added. The current `.gitignore` file:
```.venv/
__pycache__/
*.pyc
*.pyo
*.egg-info/```
 - [SkyRank/issues/4] (https://github.com/Turlure-Nael-23018992/SkyRank/issues/3): The `requirements.txt` file was removed, making `pyproject.toml` the only dependency management file.
