### 2026-01-17 #8834 @abhishektiwari
@proy-11 @Pranshu258 @MNISAR - Happy new year. I hope you're doing well. I'm reaching out once more regarding the review update as I have received no response for last 2 months. I understand that schedules can be demanding but we owe an update to authors. JOSS depends on reviewers like you, and your feedback is invaluable to both this process and the broader open source community.
