### 2026-01-28 #8834 @Pranshu258
having issues with installation:

```
chchmod +x install.sh
./install.sh
[SkyRank] Installing using CPython 3.8...
Creating virtual environment...
Installing dependencies...
Looking in links: /var/folders/f9/15sb8n7s5ps2hh5kjn36kl480000gn/T/tmpvzod5ak1
Requirement already satisfied: setuptools in ./.venv/lib/python3.8/site-packages (56.0.0)
Requirement already satisfied: pip in ./.venv/lib/python3.8/site-packages (23.0.1)
Requirement already satisfied: pip in ./.venv/lib/python3.8/site-packages (23.0.1)
Collecting pip
  Downloading pip-25.0.1-py3-none-any.whl (1.8 MB)
Installing collected packages: pip
  Attempting uninstall: pip
    Found existing installation: pip 23.0.1
    Uninstalling pip-23.0.1:
      Successfully uninstalled pip-23.0.1
Successfully installed pip-25.0.1
Obtaining file:///Users/pranshu/Documents/MyDrive/Projects/SkyRank
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Getting requirements to build editable ... done
  Preparing editable metadata (pyproject.toml) ... done
Collecting numpy==1.22.4 (from skyrank==1.0.0)
  Downloading numpy-1.22.4-cp38-cp38-macosx_11_0_arm64.whl.metadata (2.0 kB)
Collecting matplotlib==3.5.3 (from skyrank==1.0.0)
  Downloading matplotlib-3.5.3-cp38-cp38-macosx_11_0_arm64.whl.metadata (6.7 kB)
Collecting scikit-learn==1.0.2 (from skyrank==1.0.0)
  Downloading scikit_learn-1.0.2-cp38-cp38-macosx_12_0_arm64.whl.metadata (10 kB)
Collecting PyQt5==5.15.10 (from skyrank==1.0.0)
  Downloading PyQt5-5.15.10-cp37-abi3-macosx_11_0_arm64.whl.metadata (2.1 kB)
INFO: pip is looking at multiple versions of skyrank to determine which version is compatible with other requirements. This could take a while.
ERROR: Could not find a version that satisfies the requirement scipy~=1.15.3 (from skyrank) (from versions: 0.8.0, 0.9.0, 0.10.0, 0.10.1, 0.11.0, 0.12.0, 0.12.1, 0.13.0, 0.13.1, 0.13.2, 0.13.3, 0.14.0, 0.14.1, 0.15.0, 0.15.1, 0.16.0, 0.16.1, 0.17.0, 0.17.1, 0.18.0, 0.18.1, 0.19.0, 0.19.1, 1.0.0, 1.0.1, 1.1.0, 1.2.0, 1.2.1, 1.2.2, 1.2.3, 1.3.0, 1.3.1, 1.3.2, 1.3.3, 1.4.0, 1.4.1, 1.5.0, 1.5.1, 1.5.2, 1.5.3, 1.5.4, 1.6.0, 1.6.1, 1.6.2, 1.6.3, 1.7.0, 1.7.1, 1.7.2, 1.7.3, 1.8.0rc1, 1.8.0rc2, 1.8.0rc3, 1.8.0rc4, 1.8.0, 1.8.1, 1.9.0rc1, 1.9.0rc2, 1.9.0rc3, 1.9.0, 1.9.1, 1.9.2, 1.9.3, 1.10.0rc1, 1.10.0rc2, 1.10.0, 1.10.1)
ERROR: No matching distribution found for scipy~=1.15.3

ERROR: Could not find a version that satisfies the requirement humanize~=4.12.3 (from skyrank) (from versions: 0.1, 0.2, 0.3, 0.4, 0.5, 0.5.1, 1.0.0, 1.1.0, 2.0.0, 2.1.0, 2.2.0, 2.3.0, 2.4.0, 2.4.1, 2.5.0, 2.6.0, 3.0.0, 3.0.1, 3.1.0, 3.2.0, 3.3.0, 3.4.0, 3.4.1, 3.5.0, 3.6.0, 3.7.0, 3.7.1, 3.8.0, 3.9.0, 3.10.0, 3.11.0, 3.12.0, 3.13.0, 3.13.1, 3.14.0, 4.0.0, 4.1.0, 4.2.0, 4.2.1, 4.2.2, 4.2.3, 4.3.0, 4.4.0, 4.5.0, 4.6.0, 4.7.0, 4.8.0, 4.9.0, 4.10.0)
ERROR: No matching distribution found for humanize~=4.12.3

ERROR: Could not find a version that satisfies the requirement pillow~=11.2.1 (from skyrank) (from versions: 1.0, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7.0, 1.7.1, 1.7.2, 1.7.3, 1.7.4, 1.7.5, 1.7.6, 1.7.7, 1.7.8, 2.0.0, 2.1.0, 2.2.0, 2.2.1, 2.2.2, 2.3.0, 2.3.1, 2.3.2, 2.4.0, 2.5.0, 2.5.1, 2.5.2, 2.5.3, 2.6.0, 2.6.1, 2.6.2, 2.7.0, 2.8.0, 2.8.1, 2.8.2, 2.9.0, 3.0.0, 3.1.0rc1, 3.1.0, 3.1.1, 3.1.2, 3.2.0, 3.3.0, 3.3.1, 3.3.2, 3.3.3, 3.4.0, 3.4.1, 3.4.2, 4.0.0, 4.1.0, 4.1.1, 4.2.0, 4.2.1, 4.3.0, 5.0.0, 5.1.0, 5.2.0, 5.3.0, 5.4.0, 5.4.1, 6.0.0, 6.1.0, 6.2.0, 6.2.1, 6.2.2, 7.0.0, 7.1.0, 7.1.1, 7.1.2, 7.2.0, 8.0.0, 8.0.1, 8.1.0, 8.1.1, 8.1.2, 8.2.0, 8.3.0, 8.3.1, 8.3.2, 8.4.0, 9.0.0, 9.0.1, 9.1.0, 9.1.1, 9.2.0, 9.3.0, 9.4.0, 9.5.0, 10.0.0, 10.0.1, 10.1.0, 10.2.0, 10.3.0, 10.4.0)
ERROR: No matching distribution found for pillow~=11.2.1
```

Had to change the versions, after which the install script completed sucessfully:

Also, had some issue with tkinter installation because of which `skyrank-gui` isn't working. Keep getting this error -

```
skyrank-gui
Traceback (most recent call last):
  File "/Users/pranshu/Documents/MyDrive/Projects/SkyRank/.venv/bin/skyrank-gui", line 5, in <module>
    from Core.AppUIPyQT import main
  File "/Users/pranshu/Documents/MyDrive/Projects/SkyRank/Core/AppUIPyQT.py", line 18, in <module>
    from Core.Styles.StyleManager import StyleManager
  File "/Users/pranshu/Documents/MyDrive/Projects/SkyRank/Core/Styles/StyleManager.py", line 3, in <module>
    import tkinter as tk
  File "/Users/pranshu/.pyenv/versions/3.8.20/lib/python3.8/tkinter/__init__.py", line 36, in <module>
    import _tkinter # If this fails your Python may not be configured for Tk
ModuleNotFoundError: No module named '_tkinter'
```

Neither of the GUI options work for me. Only the latex command worked:

```
skyrank-graph
Welcome to the LaTeX Chart Generator!
How many algorithms do you want to compare? (1-5): 3
Do you want to use pre-existing execution data (y/n)?
y

--- Algorithm #1 ---
Enter the enum name (CoskySql, CoskyAlgorithme, RankSky, DpIdpDh, SkyIR etc.): CoskySql
✅ JSON path added: Assets/LatexData/OneAlgoData/CoskySql/ThreeColumnsData/ExecutionCoskySql369.json

--- Algorithm #2 ---
Enter the enum name (CoskySql, CoskyAlgorithme, RankSky, DpIdpDh, SkyIR etc.): CoskyAlgorithme
✅ JSON path added: Assets/LatexData/OneAlgoData/CoskyAlgo/ThreeColumnsData/ExecutionCoskyAlgo369.json

--- Algorithm #3 ---
Enter the enum name (CoskySql, CoskyAlgorithme, RankSky, DpIdpDh, SkyIR etc.): RankSky
✅ JSON path added: Assets/LatexData/OneAlgoData/RankSky/ThreeColumnsData/ExecutionRankSky369.json

Max data points to plot? Leave empty for full range:
Scale type? (LinX/LinY, LogX/LogY, LinX/LogY, etc.): LinX/LinY
✅ LaTeX chart successfully generated at: /Users/pranshu/Documents/MyDrive/Projects/SkyRank/Assets/LatexFiles/ThreeAlgosComparaison/CoskySql_CoskyAlgorithme_RankSkyLinX_LinY.tex
```

@MickaelMartinNevot can you please update the installation instructions so that these issues can be avoided. Seems like GUI options are not working on Mac (I am on MacOS Tahoe Version 26.1 (25B78))

The LaTEX version works for me. I was able to generate graphs!
