### 2026-01-28 #8834 @Pranshu258
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I would recommend that the docs be hosted on GitHub pages. Right now, they are just HTML files in the repo, which is hard to read.
