### 2026-01-28 #8834 @Pranshu258
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Can you please provide this information in the README? @MickaelMartinNevot
