### 2026-01-28 a8961f8
update files before changes for macOS installation
