### 2026-01-28 1767c83
updates file for skyrank-gui not launching on macOS
