### 2026-01-28 72a8218
Relaxed dependency versions in pyproject.toml to make installation easier on different platforms
