+++ pyproject.toml
@@ -11,15 +11,15 @@ license = { text = "MIT" }
 requires-python = ">=3.8,<3.9"
 
 dependencies = [
-  "numpy==1.22.4",
-  "matplotlib==3.5.3",
-  "scikit-learn==1.0.2",
-  "PyQt5==5.15.10",
-  "scipy~=1.15.3",
-  "colorama~=0.4.6",
-  "humanize~=4.12.3",
-  "six~=1.17.0",
-  "pillow~=11.2.1",
+  "numpy>=1.22.4",
+  "matplotlib>=3.5.3",
+  "scikit-learn>=1.0.2",
+  "PyQt5>=5.15.10",
+  "scipy>=1.15.3",
+  "colorama>=0.4.6",
+  "humanize>=4.12.3",
+  "six>=1.17.0",
+  "pillow>=11.2.1",
   "psutil",
   "GPUtil",
   "tabulate"
