+++ Core/AppUIPyQT.py
@@ -7,9 +7,6 @@ from PyQt5.QtWidgets import (
 from PyQt5.QtCore import Qt
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import StandardScaler
-from scipy.spatial import ConvexHull
 import numpy as np
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
