### 2026-02-05 #8834 @MickaelMartinNevot
Your insights and observations are highly appreciated @Pranshu258.

> Looking at the paper.md file, I see that the references section is empty. [@MickaelMartinNevot](https://github.com/MickaelMartinNevot) would you be able to add relevant citations?

The references are cited in the text throughout the manuscript (e.g., Skyline operator, ranking methods, PageRank and TOPSIS), but the GitHub Markdown preview does not render the bibliography because it does not run Pandoc/citeproc. As in the `paper.md` example in the JOSS guide: [https://github.com/adrn/gala/blob/main/paper/paper.md](https://github.com/adrn/gala/blob/main/paper/paper.md).

The references are generated correctly during the JOSS PDF build from `paper.bib`file.

> having issues with installation:

The dependency versions were relaxed in the `pyproject.toml` file to facilitate installation on disparate platforms.

> Also, had some issue with tkinter installation because of which `skyrank-gui` isn't working. Keep getting this error -

This is the underlying cause of the malfunction of SkyRank-GUI. The following error is repeatedly encountered: The StyleManager import was removed from `Core/AppUIPyQT.py`. This modification prevents SkyRank GUI from importing Tkinter, which was previously identified as the root cause of the system crash on macOS.

It has been confirmed that there are no more references to `StyleManager` or `apply_styles` remaining in the PyQt application's code. It is now possible to execute skyrank-gui on macOS. However, it is probable that `skyrank-gui2` will continue to encounter failures in the absence of Tkinter. Nevertheless, the primary interface will function properly.

> I would recommend that the docs be hosted on GitHub pages. Right now, they are just HTML files in the repo, which is hard to read.

Agreed. The documentation has already been generated as static HTML, so it would be more efficient to host it on GitHub Pages from the `/docs` directory, as this would facilitate easier reading and navigation.

A dedicated "Community Guidelines" section has been incorporated into the `README` to provide a comprehensive overview of the contribution process, the reporting of issues, and the acquisition of support. This section contains links to the CONTRIBUTING file and the GitHub issue tracker.
