### 2026-02-05 36eaf16
added Community Guidelines to README.rst
