### 2026-02-05 b7983e4
added github pages link to README.rst
