### 2026-02-09 #8834 @Pranshu258
Thanks for your response @MickaelMartinNevot, I have updated my review checklist accordingly.
I only have two pending items -

> Functionality: Have the functional claims of the software been confirmed?

The install script is working for me now. However, I am still facing some issues when running the program -

For example, here is a failure stack trace (it seems the result is being generated but there is some issue in subsequent steps -

```
[JSON Exporter] Exported to Assets/Export/JsonFiles/Result.json
Traceback (most recent call last):
  File "/Users/pranshu/Documents/Projects/SkyRank/Core/AppUIPyQT.py", line 297, in view_skyline_points
    pts = self.get_skyline_points()
  File "/Users/pranshu/Documents/Projects/SkyRank/Core/AppUIPyQT.py", line 239, in get_skyline_points
    if isinstance(result, dict): return [v[:3] for v in result.values()]
  File "/Users/pranshu/Documents/Projects/SkyRank/Core/AppUIPyQT.py", line 239, in <listcomp>
    if isinstance(result, dict): return [v[:3] for v in result.values()]
TypeError: 'int' object is not subscriptable
[1]    13001 abort      skyrank-gui
```

Also, I couldn't find any tests in the repository, would you be able to add appropriate tests? -
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
