### 2026-02-09 174c6cc
Fixed the issue where skyline points visualization was lagging + cleared console outputs
