### 2026-02-09 afc8fa9
refactor tests and added a main for unittests
