+++ tests/Algorithms/BBSCoskyTest.py
@@ -1,0 +1,12 @@
+import unittest
+import os
+import sys
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
+from Algorithms.BbsCosky import BbsCosky
+
+
+class TestBBSCosky(unittest.TestCase):
+    """Test suite for BBS_COSKY algorithm"""
+
+if __name__ == "__main__":
+    unittest.main()

[…]
