+++ Algorithms/Tests/BBSCoskyTest.py
@@ -1,6 +1,0 @@
-import unittest
-from Algorithms.BbsCosky import BbsCosky
-
-
-class TestBBSCosky(unittest.TestCase):
-    """Test suite for BBS_COSKY algorithm"""
