### 2026-02-16 #8834 @MickaelMartinNevot
We express our gratitude for your insightful observations, which have proven to be of substantial value @Pranshu258.

> The install script is working for me now. However, I am still facing some issues when running the program -

The TypeError has been resolved by updating `Core/AppUIPyQT.py`. The display logic was modified to correctly handle algorithms that return scores or identifiers (e.g., the "View Skyline Points" feature is operational without crashing, contingent upon the retrieval of the full point coordinates from the dataset (`SkyIR` and `DpIdpDh`).

> Also, I couldn't find any tests in the repository, would you be able to add appropriate tests? -

The existing unit tests have undergone a restructuring, with the resultant modifications being incorporated into a dedicated tests/ directory situated at the project's root directory. All tests have been updated to address import and path issues, and they can now be executed directly to verify the software's functionality.
