### 2026-02-17 #8834 @Pranshu258
@MickaelMartinNevot - thanks for fixing the installation and runtime errors. I am able to run the skyline algorithms now.

I had a question about the output graph, I was expecting a few skyline points (blue) in the graphs, but I couldn't see any. Is this expected?
