### 2026-02-17 #8834 @MickaelMartinNevot
Thanks for the feedback @Pranshu258!

With regard to the graph:

It was hypothesized that the graphs would contain several Skyline points (blue), but this hypothesis was not supported by the data. Is this occurrence within the bounds of what would be anticipated?

The issue with the graph display has been rectified. The logic was updated to ensure proper handling of disparate data dimensions (including 2D and 3D with scores) so that points are always plotted. To enhance visibility, a color update was implemented, whereby Skyline points were changed from blue to red, ensuring optimal distinction against the light gray background points. It is recommended that the reader observe the aforementioned phenomena.
