### 2026-02-18 #8834 @Pranshu258
@MickaelMartinNevot, thanks for fixing the color scheme. I suspect there is still something wrong with the scatterplot. The points don't seem to match the co-ordinates.

Also, for other algorithms, it is not clear that the user needs to provide "min,max,max" for the column comparison rule. I think providing a default input, with a description in the dialog box would be helpful. Something like this -

```
3 columns detected.
For each column, enter 'min' (lower is better) or 'max' (higher is better),
separated by commas.

Example: min,max,min

Enter 3 values:
```
