### 2026-02-19 #8834 @proy-11
> [@proy-11](https://github.com/proy-11) are you still keen to undertake review?

Hi @abhishektiwari,

Thank you for checking in. Apologies for the delay.

I've updated the checklist to reflect my progress.
I’ve opened two issues for the authors to address:

- [SkyRank/issues/6](https://github.com/Turlure-Nael-23018992/SkyRank/issues/6): Edge case: Identical points lead to inconsistent Skyline output
- [SkyRank/issues/5](https://github.com/Turlure-Nael-23018992/SkyRank/issues/5): Broken link for 'CONTRIBUTING' in README.

Thank you!
