### 2026-02-21 #8834 @MNISAR
> Hi [@MNISAR](https://github.com/MNISAR) - Checking if anything blocking you to progress with this review?
>
> Abhi

Sorry for delayed response. I have almost completed the review from my side.

Here are my concerns & issues I have created for those:
- https://github.com/Turlure-Nael-23018992/SkyRank/issues/7 (statement of need & State of the field)
- https://github.com/Turlure-Nael-23018992/SkyRank/issues/8 (adding tests)

@MickaelMartinNevot Do you mind uploading the generated PDF you mention in [this comment](https://github.com/openjournals/joss-reviews/issues/8834#issuecomment-3853972249) I am curious how the references turned out to be.
