### 2026-02-27 #8834 @MickaelMartinNevot
> Thanks [@Pranshu258](https://github.com/Pranshu258) driving excellent review from your end on this submission.
>
> [@MickaelMartinNevot](https://github.com/MickaelMartinNevot) Seems like software has a lot of paper cuts. I am concerned that unless documentation is improved, normal user will face uphill battle here.

We would like to express our gratitude for this feedback @abhishektiwari .

It is acknowledged that documentation is paramount for usability. In response, a comprehensive revision and expansion of the documentation has been undertaken to mitigate challenges experienced by new users. Specifically, the following elements are in place:
 - The installation instructions have been simplified and clarified.
 - A minimal working example has been incorporated to illustrate a comprehensive fundamental workflow.
 - The usage documentation was expanded to provide a more thorough exposition of the fundamental concepts and the anticipated inputs and outputs.
 - The inline documentation and docstrings have been enhanced.
 - The document has been updated to include additional guidance on common failure modes and troubleshooting procedures.

In a very specific sense, with regard to the realm of software documentation and the assessment of usability: to address these challenges, a series of modifications have been implemented with the objective of mitigating "paper cuts" and enhancing the software's ease of use. Specifically, a Minimal Working Example (MWE) and a troubleshooting guide were incorporated into the `README`, along with new detailed documentation for data formats and the graphical interface. Additionally, the GUI was updated to incorporate clear input examples (*e.g.*, ' `min`, `max`, `min`') to assist users in avoiding errors. These updates ensure that a typical user can now set up and run the software without encountering significant issues.

The objective is to facilitate the installation and execution of a fundamental example by a novice user, thereby ensuring accessibility and ease of use. It is our belief that these revisions adequately address the concerns that have been raised; however, we remain open to the possibility of receiving further specific suggestions if any issues remain.
