### 2026-02-27 #8834 @MickaelMartinNevot
> [@MickaelMartinNevot](https://github.com/MickaelMartinNevot), thanks for fixing the color scheme. I suspect there is still something wrong with the scatterplot. The points don't seem to match the co-ordinates.
> ...
> Also, for other algorithms, it is not clear that the user needs to provide "min,max,max" for the column comparison rule. I think providing a default input, with a description in the dialog box would be helpful. Something like this -
>
> ```
> 3 columns detected.
> For each column, enter 'min' (lower is better) or 'max' (higher is better),
> separated by commas.
>
> Example: min,max,min
>
> Enter 3 values:
> ```

We would like to express our gratitude for the thorough feedback provided @Pranshu258.

In regard to the matter of the scatterplot, a comprehensive overhaul of the data unification pipeline has been executed across all algorithms. The present study proposes a novel methodology that ensures the congruence between the displayed points and their actual coordinates. This methodology involves the implementation of algorithms that operate in a consistent manner within a shared coordinate space prior to the visualization process.

In response to your recommendation, we have made the necessary modifications to the preference input dialog. The dialog now displays the number of detected columns, a description of the meanings of min and max, a concrete example that is pre-filled with the correct number of values, and a clear prompt for the expected input count.
