### 2026-02-27 #8834 @MickaelMartinNevot
Thank you, @proy-11 and @MNISAR, for your reviews and for bringing up the issues.

We have addressed all of the points you raised:
#5 (broken CONTRIBUTING link) have been resolved.
#6 (identical points edge case) thanks to your feedback, we realized that a problem could arise in certain cases involving duplicate Skyline points. Therefore, we have implemented stricter dominance rules to address this issue.
#7 (Statement of Need), the section has been updated with proper references and summaries of existing tools.
#8 (unit tests), we have added new tests as suggested and plan to expand coverage further.
