### 2026-02-27 #8834 @MickaelMartinNevot
> [@MickaelMartinNevot](https://github.com/MickaelMartinNevot) Do you mind uploading the generated PDF you mention in [this comment](https://github.com/openjournals/joss-reviews/issues/8834#issuecomment-3853972249) I am curious how the references turned out to be.

Indeed. Attached is the `paper.pdf`, which was generated by the GitHub Action and uploaded to facilitate the observation of the rendering of citations and the bibliography.

[paper.pdf](https://github.com/user-attachments/files/25605574/paper.pdf)
