### 2026-02-28 #8834 @abhishektiwari
@Pranshu258  - Checking if there is any open item in relation to last checklist item from your review.

> Functionality: Have the functional claims of the software been confirmed?
