### 2026-03-10 #8834 @MickaelMartinNevot
> [@MickaelMartinNevot](https://github.com/MickaelMartinNevot) At this point could you:
>
> * [ ]  ...
>
> Once those completed, I can then move forward with recommending acceptance of the submission.

Thank you @abhishektiwari for the post-review checklist. Everything is fine with us at this point.

* [x]  Double check Paper authors and affiliations (including ORCIDs)
* [x]  Make a tagged release of your software on Github, and list the version tag of the archived version here.
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).
* [x]  Make sure that the license listed for the archive is the same as the software license.[ ]  Please list the DOI of the archived version here.

The archived version that will be referenced in the JOSS paper is version `1.1.1`.

The DOI for the archived version is [https://zenodo.org/records/18936993](https://zenodo.org/records/18936993).
