### 2026-03-17 #8834 @abhishektiwari
@MickaelMartinNevot Please see suggested edits and fixes for the paper. Please let me know once done, I will review again. Also do check with `@editorialbot generate pdf` if references are rendered correctly.

Main Body (Edits in paper.md):

Line 24-33: I strongly suggest to move URLs in **Statement of need** section as references. These are `other commonly-used packages` and should be cited appropriately.

> To our knowledge, the only open-source visualization-focused Skyline tools are Skylens implementations such as: https://www.cg.tuwien.ac.at/courses/Visualisierung2/HallOfFame/2019/Zhao2018/html/html_doku.html, https://github.com/miguelmota/skyline / https://lab.miguelmota.com/skyline and https://github.com/github/gh-skyline. SkyLens is a visual analytics system designed specifically to explore, compare, and interpret Skyline points across multidimensional datasets [@zhaoSkyLensVisualAnalysis2017].
>
> There are a few open-source tools for queries and Skyline visualization, but none for ranking. Examples include: gkoos/skyline (https://github.com/gkoos/skyline), SkylineProf (https://skylineprof.github.io/), SkylineDet-YOLOv11Seg (https://github.com/kuazhangxiaoai/SkylineDet-YOLOv11Seg).

Line 7–8: "ranking algorithms implementations" should be "ranking algorithm implementations"
Line 19: "Interpreting Skyline remains difficult" should be "Interpreting the Skyline remains difficult"
Line 42 : "Automatic scoring using with dominance hierarchy" should be "Automatic scoring using dominance hierarchy"

References (Edits in paper.bib): Please use {} explicitly to capitalise as by default JOSS formats reference titles as sentence case.

Börzsönyi et al. (2001): "skyline" should be "Skyline"
Martin Nevot & Lakhal (2026): "skylin" should be "Skylines"
Page et al. (1998): "web" should be "Web"
Zhao et al. (2017): "skyline" should be "Skyline"
