### 2026-03-18 #8834 @MickaelMartinNevot
@abhishektiwari.

"Google’s PageRank and Beyond...": The DOI (10.1515/9781400830329) has been added to the bibliography.
"The PageRank Citation Ranking...": This is a 1998 Stanford technical report and does not have a DOI.
"Skyline ranking à la IR": it was published in the CEUR Workshop Proceedings, which do not assign DOIs.

Thank you for the suggestion. We have moved all the URLs from the "Statement of need" section to the bibliography as proper references, and updated the text to use standard citations.

Thank you for pointing out these typos and capitalization issues. We have applied all the requested corrections.

In `paper.md`, we corrected the phrasing in the introduction and feature list (*e.g.*, "ranking algorithm", "the Skyline", and "using dominance").

In `paper.bib`, we added explicit braces `{}` to preserve the required capitalization for the references (Börzsönyi et al., Martin Nevot & Lakhal, Page et al., and Zhao et al.), and we changed "skylin" to "Skylines" as requested.

All files have been updated accordingly.
