### 2026-03-19 #8834 @abhishektiwari
@MickaelMartinNevot Thanks for updating the paper. I see some inconstancies wrt to use of oxford comma,

E.g. line 10, 13-14, 17-18, 22-23, etc. use oxford comma but then missing in title, on Line 6, 8, etc. Can you update the paper to follow one style.
