### 2026-03-19 #8834 @MickaelMartinNevot
@abhishektiwari, we updated the paper to use the Oxford comma consistently in the title and in all lists of three or more items. This ensures a uniform writing style throughout the document.

The version tag has been updated to v1.1.1 in the project metadata, and the correct Zenodo archive DOI (`10.5281/zenodo.18936993`) has been added to the repository (`README and `pyproject.toml` files).
