### 2026-03-19 #8834 @abhishektiwari
@MickaelMartinNevot There is another minor issue causing error with XML metadata generation. Duplicate label `mylabel` for images in `paper.md`. See below, can you make sure image label are distinct e.g., use `mylabel1` and `mylabel2` or just remove the label if not needed or used.

```
![Graphical interface with Skyline points visualization.\label{mylabel}](SkyrankUI.png)

![Latex graph generation.\label{mylabel}](SkyrankLatex.png)
```
