### 2026-03-20 #8834 @MickaelMartinNevot
@abhishektiwari, this assertion is comprehensible. The image captions have been modified to accommodate the rectification of an error in the XML.
