### 2026-03-24 #8834 @crvernon
@MickaelMartinNevot - I just need you to address the following before I accept for publication:

In your paper:

- [x] Concerning the citations with no dates listed (e.g., Mota, n.d.).  Either add a accessed date, date of formal release if an accompanying GitHub release exists to the bib record so a date is present in these citations.  Otherwise, simply link to the discussed website in the paper.  I would prefer a citation with a date/date accessed if possible.
- [x] There is no need for the bold font throughout the text to provide emphasis.  You can reformat the bold text in the paragraphs to normal.
- [x] LINE 111:  The "a" in "athens", the "g" in "greece", and the "m" in "march" should be capitalized.  You can do this in your bib file by placing characters you wish to maintain formatting of in curly brackets.

Let me know when these changes have been made and I'll accept after confirmation.  Thank you.
