### 2026-03-24 #8834 @MickaelMartinNevot
> [@MickaelMartinNevot](https://github.com/MickaelMartinNevot) - I just need you to address the following before I accept for publication:
> ...
>
> Let me know when these changes have been made and I'll accept after confirmation. Thank you.

@crvernon Thank you for your feedback. It is greatly appreciated!

Regarding the requested improvements:
* [x] Concerning citations with no dates, we have updated the `paper.bib` file. We added the release or last update year, as well as an access date (urldate: 2026-03-24), to all relevant records.
* [x] There is no need for bold font; we have removed all bold formatting used for emphasis throughout the paragraphs in `paper.md` and replaced it with regular text.
* [x] LINE 111:  We corrected the capitalization of "Athens," "Greece," and "March" in the `paper.bib` file by enclosing them in curly brackets to maintain the correct formatting.
