### 2026-03-25 #8834 @crvernon
🥳 Congratulations on your new publication @MickaelMartinNevot ! Many thanks to @abhishektiwari for editing and @proy-11, @Pranshu258, and @MNISAR for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
