### 2026-03-25 #8834 @MickaelMartinNevot
> 🥳 Congratulations on your new publication [@MickaelMartinNevot](https://github.com/MickaelMartinNevot) ! Many thanks to [@abhishektiwari](https://github.com/abhishektiwari) for editing and [@proy-11](https://github.com/proy-11), [@Pranshu258](https://github.com/Pranshu258), and [@MNISAR](https://github.com/MNISAR) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
>
> Please consider becoming a reviewer for JOSS if you are not already: https://reviewers.joss.theoj.org/join

Thank you, everyone! @crvernon My last name is Martin Nevot. It's written correclty in the `paper.md` file, but in some places, it seems to be written as only Nevot with Martin as second name.  I hope that's okay.
